js: Move region and labels into Graph component (#1310)
